requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
